Cascaded feedback implementation
