+++ modules/auth/supabase/forms/confirm-reset-password.form.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from "next-intl";
 import { yupResolver } from "@hookform/resolvers/yup";
 import AuthService from "@/services/auth/auth-supabase.service";
 import form_style from "./styles/form.module.css";
-import { confirm_reset_password_schema } from "./schemas/auth-schema";
+import { confirm_reset_password_schema } from "./auth.schema";
 import messageHandler from "@/core/message-handler";
 import { getStaticPropsWithTranslations } from "@/modules/lang/props";
 import { GetStaticProps } from "next";
